@@ -70,15 +60,17 @@
     "karma-phantomjs-launcher": "^0.2.0",
     "karma-sourcemap-loader": "^0.3.5",
     "karma-webpack": "^1.7.0",
+    "object-assign": "^4.0.1",
     "phantomjs": "^1.9.17",
     "protractor": "^2.1.0",
     "raw-loader": "^0.5.1",
+    "rimraf": "^2.4.3",
     "style-loader": "^0.12.2",
     "tsd": "^0.6.4",
-    "typescript": "^1.5.3",
-    "typescript-simple-loader": "^0.3.2",
+    "typescript": "^1.6.0-beta",
+    "typescript-simple-loader": "^0.3.7",
     "url-loader": "^0.5.5",
     "webpack": "^1.10.5",
-    "webpack-dev-server": "git://github.com/ctindel/webpack-dev-server.git"
+    "webpack-dev-server": "^1.10.1"
   }
 }
