@@ -1,24 +1,49 @@
 /// <reference path="../../../typings/_custom.d.ts" />
 
 import {Component, View} from 'angular2/angular2';
-import {Http, Headers} from 'ngHttp/http';
+import {Http, Headers} from 'angular2/http';
 import {status, json} from '../../utils/fetch';
 import { Router, RouterLink } from 'angular2/router';
-
-let styles   = require('./login.css');
-let template = require('./login.html');
+import {Auth} from '../../services/auth';
 
 @Component({
   selector: 'login'
 })
 @View({
-  styles: [ styles ],
-  template: template,
-  directives: [RouterLink]
+  directives: [RouterLink],
+  styles: [ `
+    .login {
+      width: 40%;
+    }
+  `],
+  template: `
+<div class="login jumbotron center-block">
+  <h1>Login</h1>
+  <form role="form" (submit)="login($event, username.value, password.value)">
+  <div class="form-group">
+    <label for="username">Username</label>
+    <input type="text" #username class="form-control" id="username"
+placeholder="Username">
+  </div>
+  <div class="form-group">
+    <label for="password">Password</label>
+    <input type="password" #password class="form-control" id="password"
+placeholder="Password">
+  </div>
+  <button type="submit" class="btn btn-default">Submit</button>
+</form>
+</div>
+`
 })
 export class Login {
-  constructor(public router: Router, public http: Http) {
-
+  auth: Auth;
+ 
+  constructor(public router: Router, public http: Http, public auth: Auth) {
+    this.router = router;
+    this.auth = auth;
+    if (auth.isAuth()) {
+      this.router.navigate('/home');
+    }
   }
 
   login(event, username, password) {
@@ -40,7 +65,9 @@ export class Login {
       console.dir(response);
       localStorage.setItem('jwt', response.access_token);
       console.log("Login.login: response.access_token=" + response.access_token);
-      this.router.parent.navigate('/home');
+      console.log("About to navigate to /home");
+      //this.router.parent.navigate('/home');
+      this.router.navigate('/home');
     })
     .catch((error) => {
       alert(error.message);
