@@ -1,47 +1,42 @@
 /// <reference path="../../../typings/_custom.d.ts" />
 
-import {Component, View, CORE_DIRECTIVES} from 'angular2/angular2';
-import {status, text} from '../../utils/fetch';
-import {Router} from 'angular2/router';
+import {Component, View, CORE_DIRECTIVES, FORM_DIRECTIVES} from 'angular2/angular2';
+import {Route, Redirect, Router, RouterLink, ROUTER_DIRECTIVES} from 'angular2/router';
 import {Feeds} from '../../components/feeds/feeds';
-import {Http, Headers} from 'angular2/http';
 import {Auth} from '../../services/auth';
 
-let styles   = require('./home.css');
-let template = require('./home.html');
-
+function isLoggedIn() {
+    var auth = new Auth();
+    return auth.isAuth();
+}
 
 @Component({
     selector: 'home',
-    injectables: [Auth]
 })
 @View({
-  directives: [ CORE_DIRECTIVES, Feeds ],
-  styles: [ styles ],
-  template: template
+  directives: [ ROUTER_DIRECTIVES, CORE_DIRECTIVES, FORM_DIRECTIVES, Feeds ],
+  styles: [`.red {
+    color: red;
+  }`],
+  template: `
+    <div class="container">
+      <div class="row">
+        <div class=”col-md-6 col-lg-4″><feeds></feeds></div>
+        <div class=”col-md-6 col-lg-8″><feedEntries></feedEntries></div>
+      </div>
+    </div>
+    `
 })
 export class Home {
     auth: Auth;
-    user: string;
-    isAuth: boolean;
 
-    constructor(public router: Router, public http: Http, public auth: Auth) {
+    constructor(public router: Router, public auth: Auth) {
         this.router = router;
         this.auth = auth;
     
-        this.isAuth = auth.isAuth();
-    
-        if (this.isAuth) {
-            this.user = this.auth.getUser();
-        } else {
+        if (!auth.isAuth()) {
+            console.log("Home.constructor: navigating to /login");
             this.router.parent.navigate('/login');
         }
-    }
-
-    logout(event) {
-        event.preventDefault();
-        this.auth.logout();
-        this.isAuth = false;
-        this.user = null;
     }
 }
