@@ -25,6 +25,10 @@ declare module "angular2/router" {
   }
 }
 
+declare module "angular2/tools" {
+  function enableDebugTools(appRef: any): any;
+}
+
 declare module "angular2/debug" {
   var ELEMENT_PROBE_BINDINGS: any;
   class DebugElementViewListener {
@@ -53,7 +57,7 @@ declare module "angular2/pipes" {
     _observable: any;
     _updateLatestValue(value: any): any;
     _subscribe(obs: any): any;
-    transform(obs: any, args?: List<any>): any;
+    transform(obs: any, args?: Array<any>): any;
     onDestroy(): void;
   }
 }
@@ -348,7 +352,7 @@ declare module "angular2/change_detection" {
   interface Pipe {
     supports(obj: any): boolean;
     onDestroy(): void;
-    transform(value: any, args: List<any>): any;
+    transform(value: any, args: Array<any>): any;
   }
   class Pipes {
     static extend(pipes: any)
@@ -377,7 +381,7 @@ declare module "angular2/change_detection" {
 
     supports(obj: any): boolean;
     onDestroy(): void;
-    transform(value: any, args: List<any>): any;
+    transform(value: any, args: Array<any>): any;
   }
   var defaultPipeRegistry: any;
   var defaultPipes: any;
@@ -621,7 +625,8 @@ declare module "angular2/render" {
     /**
      * Calls a method on an element.
      */
-    invokeElementMethod(location: any, methodName: string, args: List<any>): void;
+    invokeElementMethod(location: any, methodName: string, args: Array<any>): void;
+
     /**
      * Sets the value of a text node.
      */
@@ -730,7 +735,7 @@ declare module "angular2/src/dom/browser_adapter" {
         attrToPropMap: any;
         query(selector: string): any;
         querySelector(el: any, selector: string): Node;
-        querySelectorAll(el: any, selector: string): List<any>;
+        querySelectorAll(el: any, selector: string): Array<any>;
         on(el: any, evt: any, listener: any): void;
         onAndCancel(el: any, evt: any, listener: any): Function;
         dispatchEvent(el: any, evt: any): void;
@@ -745,8 +750,8 @@ declare module "angular2/src/dom/browser_adapter" {
         firstChild(el: any): Node;
         nextSibling(el: any): Node;
         parentElement(el: any): any;
-        childNodes(el: any): List<Node>;
-        childNodesAsList(el: any): List<any>;
+        childNodes(el: any): Array<Node>;
+        childNodesAsList(el: any): Array<any>;
         clearNodes(el: any): void;
         appendChild(el: any, node: any): void;
         removeChild(el: any, node: any): void;
@@ -774,7 +779,7 @@ declare module "angular2/src/dom/browser_adapter" {
         hasProperty(element: any, name: string): boolean;
         getElementsByClassName(element: any, name: string): any;
         getElementsByTagName(element: any, name: string): any;
-        classList(element: any): List<any>;
+        classList(element: any): Array<any>;
         addClass(element: any, classname: string): void;
         removeClass(element: any, classname: string): void;
         hasClass(element: any, classname: string): any;
@@ -821,7 +826,7 @@ declare module "angular2/src/dom/dom_adapter" {
         attrToPropMap: any;
         query(selector: string): any;
         querySelector(el: any, selector: string): Node;
-        querySelectorAll(el: any, selector: string): List<any>;
+        querySelectorAll(el: any, selector: string): Array<any>;
         on(el: any, evt: any, listener: any): void;
         onAndCancel(el: any, evt: any, listener: any): Function;
         dispatchEvent(el: any, evt: any): void;
@@ -836,8 +841,8 @@ declare module "angular2/src/dom/dom_adapter" {
         firstChild(el: any): Node;
         nextSibling(el: any): Node;
         parentElement(el: any): any;
-        childNodes(el: any): List<Node>;
-        childNodesAsList(el: any): List<any>;
+        childNodes(el: any): Array<Node>;
+        childNodesAsList(el: any): Array<any>;
         clearNodes(el: any): void;
         appendChild(el: any, node: any): void;
         removeChild(el: any, node: any): void;
@@ -865,7 +870,7 @@ declare module "angular2/src/dom/dom_adapter" {
         hasProperty(element: any, name: string): boolean;
         getElementsByClassName(element: any, name: string): any;
         getElementsByTagName(element: any, name: string): any;
-        classList(element: any): List<any>;
+        classList(element: any): Array<any>;
         addClass(element: any, classname: string): void;
         removeClass(element: any, classname: string): void;
         hasClass(element: any, classname: string): any;
@@ -912,7 +917,7 @@ declare module "angular2/src/dom/parse5_adapter" {
         attrToPropMap: any;
         query(selector: string): any;
         querySelector(el: any, selector: string): Node;
-        querySelectorAll(el: any, selector: string): List<any>;
+        querySelectorAll(el: any, selector: string): Array<any>;
         on(el: any, evt: any, listener: any): void;
         onAndCancel(el: any, evt: any, listener: any): Function;
         dispatchEvent(el: any, evt: any): void;
@@ -927,8 +932,8 @@ declare module "angular2/src/dom/parse5_adapter" {
         firstChild(el: any): Node;
         nextSibling(el: any): Node;
         parentElement(el: any): any;
-        childNodes(el: any): List<Node>;
-        childNodesAsList(el: any): List<any>;
+        childNodes(el: any): Array<Node>;
+        childNodesAsList(el: any): Array<any>;
         clearNodes(el: any): void;
         appendChild(el: any, node: any): void;
         removeChild(el: any, node: any): void;
@@ -956,7 +961,7 @@ declare module "angular2/src/dom/parse5_adapter" {
         hasProperty(element: any, name: string): boolean;
         getElementsByClassName(element: any, name: string): any;
         getElementsByTagName(element: any, name: string): any;
-        classList(element: any): List<any>;
+        classList(element: any): Array<any>;
         addClass(element: any, classname: string): void;
         removeClass(element: any, classname: string): void;
         hasClass(element: any, classname: string): any;
