@@ -2,6 +2,8 @@
 
 // Helper
 var sliceArgs = Function.prototype.call.bind(Array.prototype.slice);
+var toString = Object.prototype.toString;
+Object.assign = require('object-assign');
 var NODE_ENV = process.env.NODE_ENV || 'development';
 
 // Node
@@ -17,25 +19,41 @@ var DedupePlugin   = webpack.optimize.DedupePlugin;
 var DefinePlugin   = webpack.DefinePlugin;
 var BannerPlugin   = webpack.BannerPlugin;
 
-var BOWER_DIR = __dirname + '/bower_components';
 
 /*
  * Config
  */
 
-var config = {
-  devtool: 'source-map',
-  // devtool: 'eval',
+module.exports = {
+  devtool: env({
+    'development': 'eval',
+    'all': 'source-map'
+  }),
 
-  debug: true,
-  cache: true,
-  // our Development Server configs
+  debug: env({
+    'development': true,
+    'all': false
+  }),
+  cache: env({
+    // 'development': false
+    'all': true
+  }),
+  verbose: true,
+  displayErrorDetails: true,
+  context: __dirname,
+  stats: env({
+    'all': {
+      colors: true,
+      reasons: true
+    }
+  }),
+
+  // our Development Server config
   devServer: {
     inline: true,
     colors: true,
     historyApiFallback: true,
     contentBase: 'src/public',
-    //contentBase: '.',
     publicPath: '/__build__',
     staticRoutes: [ { 'route': '/scripts', 'dir' : __dirname + '/node_modules'} ]
   },
@@ -47,8 +65,7 @@ var config = {
       'rx',
       'zone.js',
       'reflect-metadata',
-      // 'rtts_assert/rtts_assert',
-
+      // to ensure these modules are grouped together in one file
       'angular2/angular2',
       'angular2/router',
       'angular2/http',
@@ -69,38 +86,33 @@ var config = {
   // Config for our build files
   output: {
     path: root('__build__'),
-    filename: '[name].js',
-    // filename: '[name].[hash].js',
-    sourceMapFilename: '[name].js.map',
+    filename: env({
+      'development': '[name].js',
+      'all': '[name].[hash].min.js'
+    }),
+    sourceMapFilename: env({
+      'development': '[name].js.map',
+      'all': '[name].[hash].min.js.map'
+    }),
     chunkFilename: '[id].chunk.js'
     // publicPath: 'http://mycdn.com/'
   },
 
   resolve: {
-    root: [__dirname, __dirname + 'node_modules/'],
+    root: __dirname,
     extensions: ['','.ts','.js','.json'],
     alias: {
-
       // should be angular2/http in next release
-      'angular2/http': 'node_modules/ngHttp/http.js',
+      // 'angular2/http': 'node_modules/ngHttp/http.js',
+
 
       // 'app': 'src/app',
       // 'common': 'src/common',
       // 'bindings': 'src/bindings',
-
       // 'components': 'src/app/components'
       // 'services': 'src/app/services',
       // 'stores': 'src/app/stores'
     }
-  },
-
-  /*
-   * When using `templateUrl` and `styleUrls` please use `__filename`
-   * rather than `module.id` for `moduleId` in `@View`
-   */
-  node: {
-    crypto: false,
-    __filename: true
   },
 
   module: {
@@ -115,7 +127,22 @@ var config = {
       { test: /\.html$/,  loader: 'raw' },
 
       // Support for .ts files.
-      { test: /\.ts$/,    loader: 'typescript-simple?&ignoreWarnings[]=2300&ignoreWarnings[]=2309', exclude: [
+      {
+        test: /\.ts$/,
+
+        loader: 'typescript-simple',
+
+        query: {
+          'ignoreWarnings': [
+            2300, // 2300 -> Duplicate identifier
+            2309, // 2309 -> An export assignment cannot be used in a module with other exported elements.
+            2346, // 2346 -> Supplied parameters do not match any signature of call target.
+            2432  // 2432 -> In an enum with multiple declarations, only one declaration can omit an initializer for its first enum element.
+          ]
+        },
+
+        exclude: [
+          /\.min\.js$/,
           /\.spec\.ts$/,
           /\.e2e\.ts$/,
           /web_modules/,
@@ -130,91 +157,80 @@ var config = {
     ]
   },
 
-  // plugins: plugins, // see below
-  context: __dirname,
-  stats: { colors: true, reasons: true }
+  plugins: env({
+    'production': [
+      new UglifyJsPlugin({
+        compress: {
+          warnings: false,
+          drop_debugger: env({
+            'development': false,
+            'all': true
+          })
+        },
+        output: {
+          comments: false
+        },
+        beautify: false
+      }),
+      new BannerPlugin(getBanner(), {entryOnly: true})
+    ],
+    'development': [
+      /* Dev Plugin */
+      // new webpack.HotModuleReplacementPlugin(),
+    ],
+    'all': [
+      new DefinePlugin({
+        'process.env.NODE_ENV': JSON.stringify(NODE_ENV),
+        'VERSION': JSON.stringify(pkg.version)
+      }),
+      new OccurenceOrderPlugin(),
+      new DedupePlugin(),
+
+      new CommonsChunkPlugin({
+        name: 'angular2',
+        minChunks: Infinity,
+        filename: env({
+          'development': 'angular2.js',
+          'all': 'angular2.min.js'
+        })
+      }),
+      new CommonsChunkPlugin({
+        name: 'common',
+        filename: env({
+          'development': 'common.js',
+          'all': 'common.min.js'
+        })
+      })
+    ]
+
+  }),
+
+  /*
+   * When using `templateUrl` and `styleUrls` please use `__filename`
+   * rather than `module.id` for `moduleId` in `@View`
+   */
+  node: {
+    crypto: false,
+    __filename: true
+  }
 };
 
-var commons_chunks_plugins = [
-  {
-    name: 'angular2',
-    minChunks: Infinity,
-    filename: 'angular2.js'
-  },
-  {
-    name: 'common',
-    filename: 'common.js'
+
+function env(configEnv) {
+  if (configEnv === undefined) { return configEnv; }
+
+  if (toString.call(configEnv[NODE_ENV]) === '[object Object]') {
+    return Object.assign({}, configEnv.all || {}, configEnv[NODE_ENV])
   }
-]
-
-
-//
-var environment_plugins = {
-
-  all: [
-    new DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify(NODE_ENV),
-      'VERSION': pkg.version
-    }),
-    new OccurenceOrderPlugin(),
-    new DedupePlugin(),
-  ],
-
-  production: [
-    new UglifyJsPlugin({
-      compress: {
-        warnings: false,
-        drop_debugger: false
-      },
-      output: {
-        comments: false
-      },
-      beautify: false
-    }),
-    new BannerPlugin(getBanner(), {entryOnly: true})
-  ],
-
-  development: [
-    /* Dev Plugin */
-    // new webpack.HotModuleReplacementPlugin(),
-  ]
-
-}//env
-
-
-
-
-
-
-
-
-
-if (NODE_ENV === 'production') {
-  // replace filename `.js` with `.min.js`
-  config.output.filename = config.output.filename.replace('.js', '.min.js');
-  config.output.sourceMapFilename = config.output.sourceMapFilename.replace('.js', '.min.js');
-  commons_chunks_plugins = commons_chunks_plugins.map(function(chunk) {
-    return chunk.filename.replace('.js', '.min.js');
-  });
+  else if (toString.call(configEnv[NODE_ENV]) === '[object Array]') {
+    return [].concat(configEnv.all || [], configEnv[NODE_ENV])
+  }
+  return configEnv[NODE_ENV] === undefined ? configEnv.all : configEnv[NODE_ENV];
 }
-else if (NODE_ENV === 'development') {
-  // any development actions here
-}
-
-// create CommonsChunkPlugin instance for each config
-var combine_common_chunks = commons_chunks_plugins.map(function(config) {
-  return new CommonsChunkPlugin(config);
-});
-
-// conbine everything
-config.plugins = [].concat(combine_common_chunks, environment_plugins.all, environment_plugins[NODE_ENV]);
-
-
-module.exports = config;
 
 // Helper functions
 function getBanner() {
-  return 'RSS Reader v'+ pkg.version;
+  return 'Reader v'+ pkg.version;
 }
 
 function root(args) {
